Touchscreen work completed
There is a bug about pointer of move controller that due to small pointer size sometimes it is hard to put it down after making player run.... But, a big task is completed. woah! hurray! at 2:18 am, oops! But that was a fun of finding errors, or mistakes than try to fix it. If you put yourself fully even though can't get it then know about same kind of work done by others watch and observe it and try to fit it towards your work and find your answer, solution word would be more appropriate so take the solution word